Created CoinPurse which accumulates coins and can return the total value of all coins.
Created CoinAccepter which has a method called Accept. Accept’s parameters are the physical properties of a coin, these properties are passed to the Build method on the CoinAccepter’s CoinFactory, which will return a identified coin that is added the purse. Lastly, Accept returns the total amount available in the coin purse.